Moved Middleware and Utils To Insfrastructure
