+++ backend/src/modules/reactions/routes/reactionRoutes.ts
@@ -1,6 +1,6 @@
 import { Router } from "express";
 import reactionController from "@/modules/reactions/controllers/reactionController";
-import { verifyAndGetUser } from "@/middleware/middleware";
+import { verifyAndGetUser } from "@/infrastructure/middleware/middleware";
 
 const router = Router();
