+++ backend/src/modules/visualizations/controllers/visualizationController.ts
@@ -1,7 +1,7 @@
 import { Request, Response } from "express";
 import { VisualizationService } from "@/modules/visualizations/services/visualizationService";
 import { APIResponse } from "@/types/response";
-import { sendResponse } from "@/modules/infrastructure/utilities/response";
+import { sendResponse } from "@/infrastructure/utilities/response";
 
 const visualizationService = new VisualizationService();
